API and Mood components
Added API for Angular interaction. It shouldn't be working that way but
this is only a demo so fuck it. It will contain the necessary API logic
to work with the back end, no real front end processing will be done (as
the endpoint will be an SPA)

Added Mood for mood tracking. It's a pretty simple model and should be
troublesome.